Implement the Appbar layout to contain two textviews
One textview is to show the breakdown while the other one is to show the total. When user scrolls up, the breakdown one will start to be hidden

A drawback of this implementation is I'm unable to show the fab for some reason. Need to investigate more why it's not showing up